adds DB lazy loading + /examples dir

+++ src/db/db.js
@@ -1,4 +1,6 @@
-let db = null;
+let db          = null;
+let Database    = null;
+
 
 /**
  * Initializes the database using Bun's SQLite driver when available,
@@ -9,44 +11,41 @@ let db = null;
  */
 if (typeof Bun !== "undefined") {
 
-  const { Database }    = await import("bun:sqlite");
-  db                    = new Database(process.env.SQLITE_DATABASE_PATH + 'transactions.db');
+    ({ Database } = await import("bun:sqlite"));
 } 
 else{
 
-  const Database    = (await import("better-sqlite3")).default;
-  db                = new Database(process.env.SQLITE_DATABASE_PATH + 'transactions.db');
+    Database= (await import("better-sqlite3")).default;
 }
 
 
-/**
- * Ensures the `uploads` table exists and delete entries older than 24h.
- *
- * This function must be called once before saving or fetching uploads.
- * It creates the table if it does not exist.
- *
- * @returns {void}
- */
-export function connect(){
-
-    db.prepare(`
+function initDB(){
     
-        CREATE TABLE IF NOT EXISTS uploads (
-                
-            txid        TEXT,            
-            content     TEXT,
-            file        BLOB,
-            params      TEXT,
-            created_at  DATETIME DEFAULT CURRENT_TIMESTAMP
-        )
-    `).run();
-    
-    db.prepare(`
+    if(!db){
         
-        DELETE FROM uploads
-        WHERE created_at < datetime('now', '-24 hours');
+        db = new Database(process.env.SQLITE_DATABASE_PATH + 'transactions.db');
+        
+        db.prepare(`
     
-    `).run();
+            CREATE TABLE IF NOT EXISTS uploads (
+
+                txid        TEXT PRIMARY KEY,            
+                content     TEXT,
+                file        BLOB,
+                params      TEXT,
+                created_at  DATETIME DEFAULT CURRENT_TIMESTAMP
+            )
+        `).run();
+    
+        db.prepare(`
+
+            DELETE FROM uploads
+            WHERE created_at < datetime('now', '-24 hours');
+
+        `).run();
+    }
+    
+    return db;
 }
 
 
@@ -65,20 +64,14 @@ export function connect(){
  */
 export function save(transaction_id, content, file, params){
 
-    const insert = db.prepare(`
+    const db        = initDB();    
+    const insert    = db.prepare(`
     
         INSERT OR REPLACE INTO uploads (txid, content, file, params)
         VALUES (?, ?, ?, ?)
     `);
 
-    try {
-        
-        insert.run(transaction_id, content, file, params);
-    } 
-    catch (err) {
-        
-        throw err;
-    }
+    insert.run(transaction_id, content, file, params);    
 }
 
 
@@ -90,20 +83,13 @@ export function save(transaction_id, content, file, params){
  */
 export function get(transaction_id){
 
-    const select = db.prepare(`
+    const db        = initDB();    
+    const select    = db.prepare(`
     
-        SELECT * FROM uploads
-        WHERE txid = ?
+        SELECT * FROM uploads WHERE txid = ?
     `);
 
-    try {
-        
-        return select.get(transaction_id);
-    } 
-    catch (err) {
-        
-        throw err;
-    }
+    return select.get(transaction_id);
 }
 
 
@@ -114,18 +100,11 @@ export function get(transaction_id){
  */
 export function getAll(){
 
-    const select = db.prepare(`
+    const db        = initDB();    
+    const select    = db.prepare(`
+    
         SELECT * FROM uploads        
     `);
 
-    
-    try {
-        
-        return select.all(); // returns row or undefined
-    } 
-    catch (err) {
-        
-        throw err;
-    }
+    return select.all();
 }
-
